+++ src/utils/transliterator/batak/batak-common.ts
