+++ src/utils/transliterator/batak/simalungun.ts
@@ -1,5 +1,8 @@
-import type { PlainRule, RegexRule, Rule, InputMethodEditor } from "../core"
-import { Batak } from "../batak-common";
+import { Batak } from "./batak-common";
+import type { PlainRule,
+              RegexRule,
+              Rule,
+              InputMethodEditor } from "../core";
 import { prepareRules,
          chainRule,
          ruleProduct,
@@ -9,30 +12,47 @@ import { prepareRules,
          asWordEnding
        } from "../core";
 
-const enum Toba {
-    Ta = "\u1BD7",
+const enum Simalungun {
+    A = "\u1BC1",
+    Ha = "\u1BC3",
+    Pa = "\u1BC8",
+    Wa = "\u1BCC",
+    Ra = "\u1BD3",
+    Ma = "\u1BD5",
+    Sa = "\u1BD9",
+    Ya = "\u1BDC",
+    La = "\u1BDF",
+    Nda = "\u1BE1",
+    Mba = "\u1BE3",
+
+    _i = "\u1BEB",
+    _ou = "\u1BE9\u1BE8",
+    _u = "\u1BEF",
 }
 
 const DigraphConsonants: PlainRule[] = [
+    ["m_b", Simalungun.Mba],
+    ["n_d", Simalungun.Nda],
     ["n_g", Batak.Nga],
+    ["n_y", Batak.Nya]
 ]
 
 const MonographConsonants: PlainRule[] = [
-    ["h", Batak.Ha],
-    ["k", Batak.Ha],
+    ["h", Simalungun.Ha],
+    ["k", Simalungun.Ha],
     ["b", Batak.Ba],
-    ["p", Batak.Pa],
+    ["p", Simalungun.Pa],
     ["n", Batak.Na],
-    ["w", Batak.Wa],
+    ["w", Simalungun.Wa],
     ["g", Batak.Ga],
     ["j", Batak.Ja],
     ["d", Batak.Da],
-    ["r", Batak.Ra],
-    ["m", Batak.Ma],
+    ["r", Simalungun.Ra],
+    ["m", Simalungun.Ma],
     ["t", Batak.Ta],
-    ["s", Batak.Sa],
-    ["y", Batak.Ya],
-    ["l", Batak.La],
+    ["s", Simalungun.Sa],
+    ["y", Simalungun.Ya],
+    ["l", Simalungun.La],
 ]
 
 const BatakConsonants = chainRule(
@@ -43,15 +63,21 @@ const LatinConsonants = chainRule(
     DigraphConsonants,
     MonographConsonants).map(([key, val]) => key)
 
+const InaNiSurat = BatakConsonants.concat([Batak.A])
+
 const IndependentVowels: PlainRule[] = [
     ["u", Batak.U],
-    ["o", Batak.A + Batak._o],
-    ["e", Batak.A + Batak._e],
+    ["o", Simalungun.A + Batak._o],
+    ["e", Simalungun.A + Batak._e],
     ["i", Batak.I],
-    ["a", Batak.A]
+    ["a", Simalungun.A]
 ]
 
-const DependentVowels: PlainRule[] = [
+const DigraphDependentVowels: PlainRule[] = [
+    ["ou", Simalungun._ou]
+]
+
+const MonographDependentVowels: PlainRule[] = [
     ["u", Batak._u],
     ["o", Batak._o],
     ["e", Batak._e],
@@ -59,8 +85,11 @@ const DependentVowels: PlainRule[] = [
     ["a", ""]
 ]
 
+const DependentVowels = chainRule(DigraphDependentVowels, MonographDependentVowels)
+
 const FinalConsonants: PlainRule[] = [
     ["n_g", Batak._ng],
+    ["h", Batak._h]
 ]
 
 const Punctuation: PlainRule[] = [
@@ -92,8 +121,12 @@ const ClosedConsonants: PlainRule[] =
         ClosedDigraphConsonants,
         ClosedMonographConsonants)
 
+const toDependentBatak = ([key, val]: PlainRule): RegexRule => [new RegExp(`(?<=(${InaNiSurat.join("|")}))${escape(key)}`), val]
+
+const asDependentBatak = (rules: PlainRule[]): RegexRule[] => rules.map(toDependentBatak)
+
 const FromLatinScheme: Rule[] = prepareRules(
-    chainRule(Syllables,
+    chainRule<Rule>(Syllables,
               FinalConsonants,
               ClosedConsonants,
               IndependentVowels,
@@ -106,14 +139,11 @@ const ToLatinScheme: Rule[] = prepareRules(
         asWordEnding([[Batak.Ha + Batak.Virama, "k"]]),
         asInverse(ClosedConsonants),
         asInverse(Syllables),
-        asInverse(Punctuation),
-        [[Batak.A + Batak._i, "i"],
-         [Batak.A + Batak._u, "u"]]
-    ))
+        asInverse(Punctuation)))
 
 const ReversibleLatinToLatinScheme: Rule[] =
     [
-        [new RegExp(`(?<=(${LatinConsonants.join("|")}))h`), "k"]
+        [new RegExp(`(?<=(${LatinConsonants.join("|")}))h`), "k"],
         ["n_g", "ng"]
     ]
 
@@ -134,6 +164,8 @@ const IMEScheme: Rule[] = prepareRules(chainRule<Rule>(
     makeTransitive(ClosedMonographConsonants,
                    ClosedDigraphConsonants,
                    Syllables),
+    asDependentBatak(ruleProduct(DigraphDependentVowels,
+                                 MonographDependentVowels)),
     // special rules
     [[Batak.Nga + Batak.Virama, Batak._ng], 
      [new RegExp(`${Batak.I}(${BatakConsonants.join("|")})${Batak.Virama}`),
