+++ src/utils/transliterator/batak/toba.ts
@@ -1,5 +1,8 @@
-import type { PlainRule, RegexRule, Rule, InputMethodEditor } from "../core"
-import { Batak } from "../batak-common";
+import { Batak } from "./batak-common";
+import type { PlainRule,
+              RegexRule,
+              Rule,
+              InputMethodEditor } from "../core";
 import { prepareRules,
          chainRule,
          ruleProduct,
@@ -9,47 +12,30 @@ import { prepareRules,
          asWordEnding
        } from "../core";
 
-const enum Simalungun {
-    A = "\u1BC1",
-    Ha = "\u1BC3",
-    Pa = "\u1BC8",
-    Wa = "\u1BCC",
-    Ra = "\u1BD3",
-    Ma = "\u1BD5",
-    Sa = "\u1BD9",
-    Ya = "\u1BDC",
-    La = "\u1BDF",
-    Nda = "\u1BE1",
-    Mba = "\u1BE3",
-
-    _i = "\u1BEB",
-    _ou = "\u1BE9\u1BE8",
-    _u = "\u1BEF",
+const enum Toba {
+    Ta = "\u1BD7",
 }
 
 const DigraphConsonants: PlainRule[] = [
-    ["m_b", Simalungun.Mba],
-    ["n_d", Simalungun.Nda],
     ["n_g", Batak.Nga],
-    ["n_y", Batak.Nya]
 ]
 
 const MonographConsonants: PlainRule[] = [
-    ["h", Simalungun.Ha],
-    ["k", Simalungun.Ha],
+    ["h", Batak.Ha],
+    ["k", Batak.Ha],
     ["b", Batak.Ba],
-    ["p", Simalungun.Pa],
+    ["p", Batak.Pa],
     ["n", Batak.Na],
-    ["w", Simalungun.Wa],
+    ["w", Batak.Wa],
     ["g", Batak.Ga],
     ["j", Batak.Ja],
     ["d", Batak.Da],
-    ["r", Simalungun.Ra],
-    ["m", Simalungun.Ma],
+    ["r", Batak.Ra],
+    ["m", Batak.Ma],
     ["t", Batak.Ta],
-    ["s", Simalungun.Sa],
-    ["y", Simalungun.Ya],
-    ["l", Simalungun.La],
+    ["s", Batak.Sa],
+    ["y", Batak.Ya],
+    ["l", Batak.La],
 ]
 
 const BatakConsonants = chainRule(
@@ -60,21 +46,15 @@ const LatinConsonants = chainRule(
     DigraphConsonants,
     MonographConsonants).map(([key, val]) => key)
 
-const InaNiSurat = BatakConsonants.concat([Batak.A])
-
 const IndependentVowels: PlainRule[] = [
     ["u", Batak.U],
-    ["o", Simalungun.A + Batak._o],
-    ["e", Simalungun.A + Batak._e],
+    ["o", Batak.A + Batak._o],
+    ["e", Batak.A + Batak._e],
     ["i", Batak.I],
-    ["a", Simalungun.A]
+    ["a", Batak.A]
 ]
 
-const DigraphDependentVowels: PlainRule[] = [
-    ["ou", Simalungun._ou]
-]
-
-const MonographDependentVowels: PlainRule[] = [
+const DependentVowels: PlainRule[] = [
     ["u", Batak._u],
     ["o", Batak._o],
     ["e", Batak._e],
@@ -82,11 +62,8 @@ const MonographDependentVowels: PlainRule[] = [
     ["a", ""]
 ]
 
-const DependentVowels = chainRule(DigraphDependentVowels, MonographDependentVowels)
-
 const FinalConsonants: PlainRule[] = [
     ["n_g", Batak._ng],
-    ["h", Batak._h]
 ]
 
 const Punctuation: PlainRule[] = [
@@ -118,12 +95,8 @@ const ClosedConsonants: PlainRule[] =
         ClosedDigraphConsonants,
         ClosedMonographConsonants)
 
-const toDependentBatak = ([key, val]): RegexRule => [new RegExp(`(?<=(${InaNiSurat.join("|")}))${escape(key)}`), val]
-
-const asDependentBatak = (rules: PlainRule[]): RegexRule[] => rules.map(toDependentBatak)
-
 const FromLatinScheme: Rule[] = prepareRules(
-    chainRule<Rule>(Syllables,
+    chainRule(Syllables,
               FinalConsonants,
               ClosedConsonants,
               IndependentVowels,
@@ -136,16 +109,21 @@ const ToLatinScheme: Rule[] = prepareRules(
         asWordEnding([[Batak.Ha + Batak.Virama, "k"]]),
         asInverse(ClosedConsonants),
         asInverse(Syllables),
-        asInverse(Punctuation)))
+        asInverse(Punctuation),
+        [[Batak.A + Batak._i, "i"],
+         [Batak.A + Batak._u, "u"]]
+    ))
 
 const ReversibleLatinToLatinScheme: Rule[] =
     [
-        [new RegExp(`(?<=(${LatinConsonants.join("|")}))h`), "k"]
+        [new RegExp(`(?<=(${LatinConsonants.join("|")}))h`), "k"],
         ["n_g", "ng"]
     ]
 
-export const fromLatin = (input: string): string => transliterate(input, FromLatinScheme);
-export const toLatin = (input: string): string => transliterate(input, ToLatinScheme);
+export const fromLatin = (input: string): string =>
+    transliterate(input, FromLatinScheme);
+export const toLatin = (input: string): string =>
+    transliterate(input, ToLatinScheme);
 export const toStandardLatin = (input: string): string =>
     transliterate(input, ReversibleLatinToLatinScheme)
 
@@ -161,8 +139,6 @@ const IMEScheme: Rule[] = prepareRules(chainRule<Rule>(
     makeTransitive(ClosedMonographConsonants,
                    ClosedDigraphConsonants,
                    Syllables),
-    asDependentBatak(ruleProduct(DigraphDependentVowels,
-                                 MonographDependentVowels)),
     // special rules
     [[Batak.Nga + Batak.Virama, Batak._ng], 
      [new RegExp(`${Batak.I}(${BatakConsonants.join("|")})${Batak.Virama}`),
